# docs/specs/api/openapi.yml
openapi: 3.0.3
info:
  title: Liberty Mutual Insurance Platform API
  description: |
    Java 11 enterprise insurance platform — policyholder records, policy management,
    premium recalculation, and compliance audit trail.
  version: 1.0.0
  contact:
    name: Liberty Mutual Platform Team
servers:
  - url: http://localhost:3001
    description: Customer Service
  - url: http://localhost:3002
    description: Policy Service
  - url: http://localhost:3003
    description: Audit Service
...
paths:
...
  /api/v2/audit-trail:
    get:
      tags: [AuditTrail]
      summary: Get aggregated audit trail with enrichment
      parameters:
        - name: entity_type
          in: query
          schema: { type: string }
        - name: entity_id
          in: query
          schema: { type: string, format: uuid }
        - name: source
          in: query
          schema: { type: string }
        - name: from
          in: query
          schema: { type: string, format: date-time }
        - name: to
          in: query
          schema: { type: string, format: date-time }
      responses:
        '200':
          description: Enriched audit trail
...
  /api/v2/audit-trail/{type}/{id}:
    get:
      tags: [AuditTrail]
      summary: Get audit trail for specific entity
      parameters:
        - name: type
          in: path
          required: true
          schema: { type: string }
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Entity audit trail
...